refactor(List): 抽离逻辑
